Create analyse submission file, enable easy (log and tmux) cleanup of Azure machine through cleanup.py, improvements to logger (various folders, current/latest.log, etc.), alteration to read in zipped pickles to get unbalanced training/test set, change to data class to train with unbalanced training set, change to gconv wrapper to allow various input/output channel combinations, 2D and 3D augmentation of batch, subclassing architectures in models3d to ensure both Z3 and G variants have same architecture

+++ code/cleanup.py
@@ -11,9 +11,11 @@ keep_limit = 5
 def cleanup_all():
     # clean up current log files
     logfiles = glob.glob(logdir + '/current*log')
-    for logfile in logfiles:
-        print('Removing %s.' % logfile)
-        os.remove(logfile)
+    if len(logfiles) > 0:
+        print('Cleaning up current log files: ')
+        for logfile in logfiles:
+            print('\t -%s.' % logfile)
+            os.remove(logfile)
 
     # clean up folders
     for data_folder in folders:
@@ -24,11 +26,13 @@ def cleanup_all():
             logfiles.sort()
 
             # remove old files
+            print('\nRemoving old files in %s: ' % data_folder)
             for _, logfile in logfiles[:keep_limit + 1]:
-                print('Removing %s.' % logfile)
+                print('\t -%s.' % logfile)
                 os.remove(logfile)
 
             # rename kept log files
+            print('\nRenaming kept log files in %s' % data_folder)
             for i, (_, logfile) in enumerate(logfiles[keep_limit + 1:]):
                 folder_structure = logfile.split('/')
                 fname_split = folder_structure[-1].split('_')
@@ -36,14 +40,18 @@ def cleanup_all():
                 new_fname = '_'.join(fname_split)
                 folder_structure[-1] = new_fname
                 dst = '/'.join(folder_structure)
-                print('Renaming %s to %s' % (logfile, dst))
+                print('\t - %s to %s' % (logfile, dst))
                 os.rename(logfile, dst)
+
+    # close all tmux windows
+    try:
+        os.system('tmux kill-server')
+    except:
+        raise Exception
 
 def cleanup_broken():
     broken_logfiles = glob.glob(logdir + '/*/broken*')
-    if len(broken_logfiles) == 0:
-        print('No broken logfiles.')
-    else:
+    if len(broken_logfiles) != 0:
         for logfile in broken_logfiles:
             print('Removing %s.' % logfile)
             os.remove(logfile)
@@ -52,10 +60,12 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--keep', type=int, default=5)
     parser.add_argument('--broken', action='store_true')
+
     args = parser.parse_args()
 
     if args.broken:
         cleanup_broken()
     else:
+        cleanup_broken()
         keep_limit = args.keep
         cleanup_all()
